refactor: move `generate_trtllm_engine_config` into `config_gen`
