+++ src/ditto/__init__.py
@@ -1,6 +1,6 @@
 # Note: logging must be imported first to set up the loguru logging.
 from . import logging, constants  # noqa: I001
-from .config_gen import PretrainedConfigGenerationError, generate_trtllm_pretrained_config
+from .config_gen import PretrainedConfigGenerationError, generate_trtllm_engine_config
 from .convert import convert
 from .export import export
 from .inline import inline
